Reorganize models and views in different files
Split one model per file, as advised in the contribution guide.
