+++ queue_job/__manifest__.py
@@ -14,6 +14,11 @@
         "security/security.xml",
         "security/ir.model.access.csv",
         "views/queue_job_views.xml",
+        "views/queue_job_channel_views.xml",
+        "views/queue_job_function_views.xml",
+        "wizards/queue_jobs_to_done_views.xml",
+        "wizards/queue_requeue_job_views.xml",
+        "views/queue_job_menus.xml",
         "data/queue_data.xml",
         "data/queue_job_function_data.xml",
     ],
