+++ queue_job/models/__init__.py
@@ -1,3 +1,5 @@
 from . import base
 from . import ir_model_fields
 from . import queue_job
+from . import queue_job_channel
+from . import queue_job_function
