+++ queue_job/views/queue_job_menus.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<odoo>
+
+    <menuitem
+        id="menu_queue_job_root"
+        name="Job Queue"
+        web_icon="queue_job,static/description/icon.png"
+        groups="group_queue_job_manager"
+    />
+
+    <menuitem id="menu_queue" name="Queue" parent="menu_queue_job_root" />
+
+    <menuitem
+        id="menu_queue_job"
+        action="action_queue_job"
+        sequence="10"
+        parent="menu_queue"
+    />
+
+    <menuitem
+        id="menu_queue_job_channel"
+        action="action_queue_job_channel"
+        sequence="12"
+        parent="menu_queue"
+    />
+
+    <menuitem
+        id="menu_queue_job_function"
+        action="action_queue_job_function"
+        sequence="14"
+        parent="menu_queue"
+    />
+
+</odoo>
