+++ queue_job/wizards/__init__.py
@@ -1,0 +1,2 @@
+from . import queue_requeue_job
+from . import queue_jobs_to_done
